colpali_engine/vision/image_processor.py
"""
Image processing utilities for document frames.

This module provides image standardization, preprocessing, and optimization
utilities for preparing document images for ColPali processing.
"""

import logging
from typing import List, Optional, Tuple
from PIL import Image

logger = logging.getLogger(__name__)


class ImageProcessor:
    """
    Image processing utilities for document frames.

    Handles image standardization, resolution normalization, and quality
    optimization for vision model input.
    """

    def __init__(self, target_dpi: int = 300, target_format: str = "RGB"):
        """
        Initialize image processor.

        Args:
            target_dpi: Target resolution for processed images
            target_format: Target color format
        """
        self.target_dpi = target_dpi
        self.target_format = target_format

        logger.info(f"ImageProcessor initialized: {target_dpi} DPI, {target_format} format")

    def standardize_images(
        self,
        images: List[Image.Image],
        preserve_aspect_ratio: bool = True
    ) -> List[Image.Image]:
        """
        Standardize image frames for consistent processing.

        This will be implemented in COLPALI-202.

        Args:
            images: List of PIL Image objects
            preserve_aspect_ratio: Whether to preserve aspect ratios

        Returns:
            List of standardized images
        """
        logger.info(f"Standardizing {len(images)} images - TODO: Implementation needed")
        # TODO: Implement resolution normalization
        # TODO: Implement color space conversion
        # TODO: Implement quality optimization
        # TODO: Add metadata preservation

        return images

    def optimize_for_colpali(self, image: Image.Image) -> Image.Image:
        """
        Optimize image for ColPali model input requirements.

        Args:
            image: PIL Image to optimize

        Returns:
            Optimized image
        """
        logger.info("Optimizing image for ColPali - TODO: Implementation needed")
        # TODO: Implement ColPali-specific optimizations
        return image

    def extract_patches(
        self,
        image: Image.Image,
        patch_size: Tuple[int, int] = (32, 32)
    ) -> List[Tuple[Image.Image, Tuple[int, int]]]:
        """
        Extract patches from image with coordinate tracking.

        Args:
            image: PIL Image to extract patches from
            patch_size: Size of patches to extract

        Returns:
            List of (patch_image, coordinates) tuples
        """
        logger.info("Extracting patches - TODO: Implementation needed")
        # TODO: Implement patch extraction
        # TODO: Add coordinate tracking
        return []